Basic webapp configuration
Includes a custom compilation system because I was too lazy to find
another one. The app can be compiled with the command
`npm run build` after creating the `build` directory for the output.

The web code itself implements the basic features of the
texthooker, which I've had working for a while. This includes
arrangement of inserted text, tracking of line and character
count, line deletion, and the option to switch between upwards
and downwards flow of lines.